Refactor review files and diff
